chore: document type update, cleanup and restructuring of files
