+++ packages/content/src/atoms/desing-library-bridge/constants.ts
@@ -1,0 +1,18 @@
+/**
+ * Event name for component preview updates from design library
+ * @internal
+ */
+export const DESIGN_LIBRARY_COMPONENT_PREVIEW_EVENT_NAME = 'component:atoms:preview';
+
+/**
+ * Event name for the atoms catalog registration (replaces legacy `atom:registry`).
+ * @internal
+ */
+export const DESIGN_LIBRARY_ATOMS_CATALOG_EVENT_NAME = 'atoms:catalog';
+
+/**
+ * Event to send to design library when rendering atoms error occurs
+ * @internal
+ */
+export const DESIGN_LIBRARY_ATOMS_ERROR_EVENT_NAME = 'atoms:error';
+
