+++ packages/content/src/atoms/index.ts
@@ -1,14 +1,18 @@
-/**
- * Component Layout document types, guards, and utilities exposed on the
- * `@sitecore-content-sdk/content/atoms` entry point.
- *
- * The document model is aligned with json-render's flat spec format.
- */
-
-export { type Document } from './component-layout/document';
+// export { type Document } from './component-layout/document';
 export type {
   SerializedCatalog,
   SitecoreComponentMeta,
   AtomCatalogComponentEntry,
   AtomCatalogActionEntry,
+  Document,
 } from './types';
+
+export {
+  AtomCatalogEntry,
+  ActionCatalogEntry,
+  AtomsCatalogPayload,
+  getDesignLibraryAtomsCatalogEvent,
+  sendAtomsErrorEvent,
+  DesignLibraryAtomsError,
+  addDocumentUpdateHandler,
+} from './desing-library-bridge';
