--- packages/content/src/editing/atoms-builder/index.ts
@@ -1,10 +1,0 @@
-export {
-  AtomCatalogEntry,
-  ActionCatalogEntry,
-  AtomsCatalogPayload,
-  getDesignLibraryAtomsCatalogEvent,
-  getDesignLibraryAtomsErrorEvent,
-  sendAtomsErrorEvent,
-  DesignLibraryAtomsError,
-  addDocumentUpdateHandler,
-} from './atoms-builder';
